refactor Long to Int add shadow for imageview and fix bindingadapters

+++ app/src/main/java/fr/im/salimi/projectmanager/data/helpers/Address.kt
@@ -1,9 +1,9 @@
 package fr.im.salimi.projectmanager.data.helpers
 
 class Address(
-    var roadNumber: Long = 0,
+    var roadNumber: Int = 0,
     var roadName: String = "",
     var city: String = "",
-    var postalCode: Long = 0,
+    var postalCode: Int = 0,
     var country: String = ""
 )
